# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de troc de services entre particuliers, basé sur des crédits-temps.",
        "title": "Barterswap",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/exchanges": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Lister mes échanges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrer par statut",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/exchange.Exchange"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Créer une demande d'échange",
                "parameters": [
                    {
                        "description": "service_id",
                        "name": "exchange",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/exchange.Exchange"
                        }
                    }
                }
            }
        },
        "/exchanges/{id}": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Récupérer un échange par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID échange",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exchange.Exchange"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exchanges/{id}/accept": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Accepter un échange",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID échange",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exchanges/{id}/cancel": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Annuler un échange",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID échange",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exchanges/{id}/complete": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Terminer un échange accepté",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID échange",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exchanges/{id}/reject": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Refuser un échange",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID échange",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "exchange.Exchange": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "owner_id": {
                    "type": "integer"
                },
                "requester_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "status": {
                    "description": "pending, accepted, rejected, cancelled, completed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "X-UserID",
            "in": "header"
        }
    }
}